migrate other adapters to use the snowflake technique
