+++ plugins/snowflake/dbt/adapters/snowflake/impl.py
@@ -1,9 +1,7 @@
-from concurrent.futures import ThreadPoolExecutor, as_completed
-from typing import Mapping, Any, Optional, Set
+from typing import Mapping, Any, Optional
 
 import agate
 
-from dbt.adapters.base.relation import InformationSchema
 from dbt.adapters.sql import SQLAdapter
 from dbt.adapters.snowflake import SnowflakeConnectionManager
 from dbt.adapters.snowflake import SnowflakeRelation
@@ -11,9 +9,6 @@ from dbt.adapters.snowflake import SnowflakeColumn
 from dbt.contracts.graph.manifest import Manifest
 from dbt.exceptions import RuntimeException
 from dbt.utils import filter_null_values
-
-
-GET_CATALOG_MACRO_NAME = 'snowflake_get_catalog'
 
 
 class SnowflakeAdapter(SQLAdapter):
@@ -31,7 +26,9 @@ class SnowflakeAdapter(SQLAdapter):
         return "CURRENT_TIMESTAMP()"
 
     @classmethod
-    def _catalog_filter_table(cls, table, manifest):
+    def _catalog_filter_table(
+        cls, table: agate.Table, manifest: Manifest
+    ) -> agate.Table:
         # On snowflake, users can set QUOTED_IDENTIFIERS_IGNORE_CASE, so force
         # the column names to their lowercased forms.
         lowered = table.rename(
@@ -78,52 +75,6 @@ class SnowflakeAdapter(SQLAdapter):
         previous = self._get_warehouse()
         self._use_warehouse(warehouse)
         return previous
-
-    def _get_one_catalog(
-        self,
-        information_schema: InformationSchema,
-        schemas: Set[str],
-        manifest: Manifest,
-    ) -> agate.Table:
-
-        name = '.'.join([
-            str(information_schema.database),
-            'information_schema'
-        ])
-
-        # calculate the possible schemas for a given schema name
-        all_schema_names: Set[str] = set()
-        for schema in schemas:
-            all_schema_names.update({schema, schema.lower(), schema.upper()})
-
-        with self.connection_named(name):
-            kwargs = {
-                'information_schema': information_schema,
-                'schemas': all_schema_names
-            }
-            table = self.execute_macro(GET_CATALOG_MACRO_NAME,
-                                       kwargs=kwargs,
-                                       release=True)
-
-        results = self._catalog_filter_table(table, manifest)
-        return results
-
-    def get_catalog(self, manifest: Manifest) -> agate.Table:
-        # snowflake is super slow. split it out into the specified threads
-        num_threads = self.config.threads
-        schema_map = self._get_cache_schemas(manifest)
-        catalogs: agate.Table = agate.Table(rows=[])
-
-        with ThreadPoolExecutor(max_workers=num_threads) as executor:
-            futures = [
-                executor.submit(self._get_one_catalog, info, schemas, manifest)
-                for info, schemas in schema_map.items() if len(schemas) > 0
-            ]
-            for future in as_completed(futures):
-                catalog = future.result()
-                catalogs = agate.Table.merge([catalogs, catalog])
-
-        return catalogs
 
     def post_model_hook(
         self, config: Mapping[str, Any], context: Optional[str]
