+++ test/unit/test_postgres_adapter.py
@@ -180,7 +180,8 @@ class TestPostgresAdapter(unittest.TestCase):
             connect_timeout=10)
 
     @mock.patch.object(PostgresAdapter, 'execute_macro')
-    def test_get_catalog_various_schemas(self, mock_execute):
+    @mock.patch.object(PostgresAdapter, '_get_cache_schemas')
+    def test_get_catalog_various_schemas(self, mock_get_schemas, mock_execute):
         column_names = ['table_database', 'table_schema', 'table_name']
         rows = [
             ('dbt', 'foo', 'bar'),
@@ -191,6 +192,8 @@ class TestPostgresAdapter(unittest.TestCase):
         ]
         mock_execute.return_value = agate.Table(rows=rows,
                                                 column_names=column_names)
+
+        mock_get_schemas.return_value.items.return_value = [(mock.MagicMock(database='dbt'), {'foo', 'FOO', 'quux'})]
 
         mock_manifest = mock.MagicMock()
         mock_manifest.get_used_schemas.return_value = {('dbt', 'foo'),
